Rename version file for version substitution
