+++ dataclasses_json/__init__.py
@@ -5,7 +5,7 @@ from dataclasses_json.cfg import (config, global_config,
                                   Exclude, LetterCase)
 from dataclasses_json.undefined import CatchAll, Undefined
 
-from dataclasses_json._version import __version__
+from dataclasses_json.__version__ import __version__
 
 __all__ = ['DataClassJsonMixin', 'LetterCase', 'dataclass_json',
            'config', 'global_config', 'Exclude',
